backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart


